manifest transpiler to service for encryption di
